Move sections into their own folder
