+++ spec/site_prism/page_objects/page/organisations/course_fees.rb
@@ -15,7 +15,7 @@ module PageObjects
         element :flash, '.govuk-success-summary'
         element :warning_message, '[data-copy-course=warning]'
 
-        section :copy_content, CopyContentSection
+        section :copy_content, PageObjects::Section::CopyContentSection
       end
     end
   end
