+++ spec/site_prism/page_objects/page/organisations/course_salary.rb
@@ -12,7 +12,7 @@ module PageObjects
         element :flash, '.govuk-success-summary'
         element :warning_message, '[data-copy-course=warning]'
 
-        section :copy_content, CopyContentSection
+        section :copy_content, PageObjects::Section::CopyContentSection
       end
     end
   end
